refactor: Split Bookings page components into different files.
