+++ pages/grid.tsx
@@ -1,14 +1,11 @@
-import * as React from 'react'
-import { useState, useRef, useEffect, useContext } from 'react'
-import { motion, AnimatePresence } from 'framer-motion'
+import * as React from 'react';
+import { useState, useEffect } from 'react';
+import { motion } from 'framer-motion';
 
 import {
   Box,
   HStack,
   VStack,
-  Center,
-  Text,
-  Select,
   Flex,
   Tab,
   TabList,
@@ -23,171 +20,27 @@ import {
   Button,
   MenuOptionGroup,
 } from '@chakra-ui/react';
-import { useBoolean, useDisclosure } from '@chakra-ui/react';
+import { useDisclosure } from '@chakra-ui/react';
 
 import eachMinuteOfInterval from 'date-fns/eachMinuteOfInterval';
-import format from 'date-fns/format';
 
-import { BookingConfirmationPopup } from '../components/BookingConfirmationPopup';
+import { BookingConfirmationPopup } from '../components/booking/BookingConfirmationPopup';
 import { BookingsContext, BookingsContextValue } from './BookingsContext';
-import { isAfter, isSameDay, sub, addMinutes } from 'date-fns';
+import { sub } from 'date-fns';
 import { SingleDatepicker } from 'chakra-dayzed-datepicker';
 import Footer from '../components/Footer';
 import { NextPage } from 'next';
 import NavMenu from '../components/NavMenu';
 import { useLocalStorage } from '../components/swr-internal-state-main';
-import Calendar from '../components/Calendar'
-import { ChevronDownIcon } from '@chakra-ui/icons'
-
-// Types for the BookingsOld Components
-// To be moved to global types file after replacing the old BookingsOld page
-interface VenueBookingProps {
-  venueName: String;
-  openBookingModal: (start: Date, end: Date) => void;
-  bookingModalIsOpen: boolean;
-  timeIntervals: Date[];
-  currentVenueBookings: Array<BookingDataDisplay>;
-}
-
-interface VenueTimeCellProps {
-  onMouseDown: () => void;
-  onMouseOver: () => void;
-  booked: boolean;
-  selected: boolean;
-}
+import Calendar from '../components/Calendar';
+import { ChevronDownIcon } from '@chakra-ui/icons';
+import BookingsTimesCol from '../components/booking/BookingTimesCol';
+import BookingVenueCol from '../components/booking/BookingVenueCol';
 
 const BOX_HEIGHT = 8; // Ensures time labels are aligned with grid cells
 const VENUES = ['CTPH', 'Chatterbox', "Maker's Studio", 'Amphi', 'TRR', 'TRB'];
 
 const useUserInfo = () => useLocalStorage<AuthState>('token-value');
-
-// Detects clicks outside of the grid
-function useOutsideAlerter(ref: any, callback: () => void) {
-  useEffect(() => {
-    const handleClickOutside = (event: any) =>
-      ref.current && !ref.current.contains(event.target) && callback();
-
-    document.addEventListener('mousedown', handleClickOutside);
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-    };
-  }, [ref, callback]);
-}
-
-// Individual Grid Cells for the time intervals
-const VenueTimeCell: React.FC<VenueTimeCellProps> = ({
-  onMouseDown,
-  onMouseOver,
-  booked,
-  selected,
-}) => {
-  // Cell is coloured based on whether it's selected or not
-  return (
-    <Box
-      w='40'
-      h={BOX_HEIGHT}
-      bg={booked ? 'green.500' : selected ? 'blue.500' : 'gray.200'}
-      _hover={{ bg: booked ? 'green.700' : selected ? 'blue.700' : 'gray.300' }}
-      onMouseOver={onMouseOver}
-      onMouseDown={onMouseDown}
-    ></Box>
-  );
-};
-
-// Column of Time Grid Cells for a single venue
-const VenueBooking: React.FC<VenueBookingProps> = ({
-  venueName,
-  openBookingModal,
-  bookingModalIsOpen,
-  timeIntervals,
-  currentVenueBookings,
-}) => {
-  // Helper function to check if the current cell is between the first and last selected cells
-  const between = (currentIndex: number, x: number, y: number): boolean => {
-    const startIndex = Math.min(x, y);
-    const endIndex = Math.max(x, y);
-
-    return currentIndex >= startIndex && currentIndex <= endIndex;
-  };
-  // Venue column works by colouring in cells between firstSelected and lastSelected
-  // firstSelected is updated when user holds the mouse down
-  // lastSelected is updated when cursor moves over a cell while mouse is held down
-  const [mouseIsDown, setMouse] = useBoolean();
-  const [firstSelected, setFirst] = useState(-1);
-  const [lastSelected, setLast] = useState(-1);
-
-  const wrapperRef = useRef(null); //  Used to detect clicks outside of the grid
-  useOutsideAlerter(wrapperRef, () => {
-    if (bookingModalIsOpen) return; // Don't deselect if booking modal is open
-    setFirst(-1);
-    setLast(-1);
-  });
-
-  return (
-    <VStack ref={wrapperRef} spacing='0'>
-      <Text fontSize='lg'>{venueName}</Text>
-      <VStack
-        onMouseDown={setMouse.on}
-        onMouseUp={() => {
-          setMouse.off();
-          if (firstSelected === -1) return;
-          // If selection has been made, open the booking modal
-          const start = Math.min(firstSelected, lastSelected)
-          const end = Math.max(firstSelected, lastSelected)
-          openBookingModal(timeIntervals[start], addMinutes(timeIntervals[end], 30))
-        }}
-      >
-        {timeIntervals.map((el, i) => {
-          // Check if the current cell is booked
-          const isBooked = currentVenueBookings.some((booking) => {
-            return isAfter(el, booking.from) && isAfter(booking.to, el);
-          });
-          return (
-            <VenueTimeCell
-              key={i}
-              booked={isBooked}
-              onMouseDown={() => {
-                if (isBooked) return;
-                setFirst(i);
-                setLast(i);
-              }}
-              onMouseOver={() => {
-                mouseIsDown && !isBooked && setLast(i);
-              }}
-              selected={!isBooked && between(i, firstSelected, lastSelected)}
-            />
-          );
-        })}
-      </VStack>
-    </VStack>
-  );
-};
-
-// Labels for time 0000-2330
-const BookingTimes: React.FC = () => {
-  // Format the time intervals into strings
-  const timeStrings = eachMinuteOfInterval(
-    {
-      start: new Date(2000, 1, 1, 0), // Start at 0000
-      end: new Date(2000, 1, 1, 23, 59), // End at 2359
-    },
-    { step: 30 },
-  ).map((el) => format(el, 'HH:mm'));
-
-  return (
-    <VStack spacing='0'>
-      <Text fontSize={'lg'}>Time</Text>
-      <VStack>
-        {timeStrings.map((el, i) => (
-          <Center h={BOX_HEIGHT} key={el}>
-            {/* Ensures time labels are aligned with grid cells */}
-            <Box>{el}</Box>
-          </Center>
-        ))}
-      </VStack>
-    </VStack>
-  );
-};
 
 const BookingSelector: React.FC = () => {
   const { isOpen, onOpen, onClose } = useDisclosure();
@@ -216,17 +69,17 @@ const BookingSelector: React.FC = () => {
       endOfDay.setHours(23, 59, 59, 999);
       const currentBookings = await fetch(
         process.env.NEXT_PUBLIC_BACKEND_URL +
-        'bookings/all?start=' +
-        startOfDay.toISOString() +
-        '&end=' +
-        endOfDay,
+          'bookings/all?start=' +
+          startOfDay.toISOString() +
+          '&end=' +
+          endOfDay,
       );
       const allBookings = await currentBookings.json();
       console.log(allBookings);
       setAllBookings(allBookings);
     })();
     return () => {};
-  }, [startDate])
+  }, [startDate]);
 
   // Create time intervals for the current date
   const timeIntervals = (() => {
@@ -327,9 +180,9 @@ const BookingSelector: React.FC = () => {
           {/* Day view */}
           <TabPanel>
             <HStack>
-              <BookingTimes />
+              <BookingsTimesCol boxHeight={BOX_HEIGHT} />
               {VENUES.map((venueName, venueId) => (
-                <VenueBooking
+                <BookingVenueCol
                   timeIntervals={timeIntervals}
                   key={venueName}
                   venueName={venueName}
@@ -345,6 +198,7 @@ const BookingSelector: React.FC = () => {
                   }}
                   bookingModalIsOpen={isOpen}
                   currentVenueBookings={venueBookings[venueId]}
+                  boxHeight={BOX_HEIGHT}
                 />
               ))}
             </HStack>
